+++ lib/themes.dart
@@ -62,6 +62,14 @@ class Themes {
               width: 0.5,
             ),
             sidebarColor: Colors.white.withOpacity(0.4),
+            systemMessageBorder:
+                Border.all(color: const Color(0xFFD2D2D2), width: 0.5),
+            systemMessageColor: const Color(0xFFEFEFEF).withOpacity(0.95),
+            systemMessageStyle: GoogleFonts.roboto(
+              color: const Color(0xFF888888),
+              fontSize: 13,
+              fontWeight: FontWeight.w300,
+            ),
           ),
         ],
         colorScheme: colors,
@@ -277,6 +285,9 @@ class Style extends ThemeExtension<Style> {
     required this.primaryBorder,
     required this.secondaryBorder,
     required this.sidebarColor,
+    required this.systemMessageBorder,
+    required this.systemMessageColor,
+    required this.systemMessageStyle,
   });
 
   /// [Color] of the modal background barrier color.
@@ -315,6 +326,15 @@ class Style extends ThemeExtension<Style> {
   /// [Color] of the [HomeView]'s side bar.
   final Color sidebarColor;
 
+  /// [Border] to apply to system messages.
+  final Border systemMessageBorder;
+
+  /// [Color] of system messages.
+  final Color systemMessageColor;
+
+  /// [TextStyle] of system messages.
+  final TextStyle systemMessageStyle;
+
   @override
   ThemeExtension<Style> copyWith({
     Color? barrierColor,
@@ -329,6 +349,9 @@ class Style extends ThemeExtension<Style> {
     Border? primaryBorder,
     Border? secondaryBorder,
     Color? sidebarColor,
+    Border? systemMessageBorder,
+    Color? systemMessageColor,
+    TextStyle? systemMessageStyle,
   }) {
     return Style(
       barrierColor: barrierColor ?? this.barrierColor,
@@ -347,6 +370,9 @@ class Style extends ThemeExtension<Style> {
       primaryBorder: primaryBorder ?? this.primaryBorder,
       secondaryBorder: secondaryBorder ?? this.secondaryBorder,
       sidebarColor: sidebarColor ?? this.sidebarColor,
+      systemMessageBorder: systemMessageBorder ?? this.systemMessageBorder,
+      systemMessageColor: systemMessageColor ?? this.systemMessageColor,
+      systemMessageStyle: systemMessageStyle ?? this.systemMessageStyle,
     );
   }
 
@@ -383,6 +409,15 @@ class Style extends ThemeExtension<Style> {
       primaryBorder: Border.lerp(primaryBorder, other.primaryBorder, t)!,
       secondaryBorder: Border.lerp(secondaryBorder, other.secondaryBorder, t)!,
       sidebarColor: Color.lerp(sidebarColor, other.sidebarColor, t)!,
+      systemMessageBorder:
+          Border.lerp(systemMessageBorder, other.systemMessageBorder, t)!,
+      systemMessageColor:
+          Color.lerp(systemMessageColor, other.systemMessageColor, t)!,
+      systemMessageStyle: TextStyle.lerp(
+        systemMessageStyle,
+        other.systemMessageStyle,
+        t,
+      )!,
     );
   }
 }