+++ CHANGELOG.md
@@ -16,6 +16,14 @@ All user visible changes to this project will be documented in this file. This p
 - UI:
     - Media panel:
         - Dock buttons persistence. ([#159], [#137])
+    - Chat info page:
+        - Chat avatar uploading and deleting. ([#163], [#146])
+
+### Changed
+
+- UI:
+    - Chat page:
+        - Redesigned system messages. ([#161], [#158])
 
 ### Changed
 
@@ -24,9 +32,13 @@ All user visible changes to this project will be documented in this file. This p
         - Redesigned messages and forwards. ([#162], [#151])
 
 [#137]: /../../issues/137
+[#146]: /../../issues/146
 [#151]: /../../issues/151
+[#158]: /../../issues/158
 [#159]: /../../pull/159
+[#161]: /../../pull/161
 [#162]: /../../pull/162
+[#163]: /../../pull/163